Update tenzing.py - fixing website duplicates
I'm trying to ensure that if a contributor has their ORCID on one Tenzing but not another, the code will still treat then as a single contributor and there will only be one entry on the website. At the moment, if a person has an ORCID on one Tenzing but not another they get two separate entries on the website.

So what I have tried to do is get it to group by name and then if there's an ORCID iD, to propagate that across all contributions for that contributor. 

It worked on a test df, but I'm out of my depth here, so if it's still way off just let me know and I'll go back to the drawing board!